editoast: isolate runserver CLI into its own module
